move archiveMessages into helpers file
